change count down method

+++ models/item.js
@@ -190,15 +190,18 @@ const deleteItem = async function (id) {
 };
 
 const countdownDeleteItem = function (item) {
-  const time =
-    (item.countdown.day * 24 * 60 * 60 +
-      item.countdown.hour * 60 * 60 +
-      item.countdown.minute * 60 +
-      item.countdown.second) *
-    1000;
-  setTimeout(function () {
+  // const time =
+  //   (item.countdown.day * 24 * 60 * 60 +
+  //     item.countdown.hour * 60 * 60 +
+  //     item.countdown.minute * 60 +
+  //     item.countdown.second) *
+  //   1000;
+  // setTimeout(function () {
+  //   deleteItem(item._id);
+  // }, time);
+  if(item.timeLeft.day === 0 && item.timeLeft.hour === 0 && item.timeLeft.minute === 0 && item.timeLeft.second === 0){
     deleteItem(item._id);
-  }, time);
+  }
 };
 
 const calculateTimeLeft = function (item) {
